requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.83
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
decorator==5.1.1
distlib==0.3.6
einops==0.2.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.2.3
flask-babel==3.1.0
Flask-Mail==0.9.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
gdown==4.4.0
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
gradio==3.11.0
greenlet==2.0.2
grpcio==1.54.2
gTTS==2.2.4
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.24.0
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
linkify-it-py==2.0.2
lmdb==1.4.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdit-py-plugins==0.3.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
namegenerator==1.0.6
networkx==2.6.3
numpy==1.21.6
onnx==1.14.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.12
packaging==21.3
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pafy==0.5.5
pandas==1.3.5
paramiko==3.1.0
pdf2docx==0.5.6
pefile==2022.5.30
Pillow==9.3.0
platformdirs==3.2.0
playsound==1.2.2
prefetch-generator==1.0.1
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.4
pyarrow==11.0.0
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.1.2
Pygments==2.15.1
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
Pympler==1.0.1
PyMuPDF==1.20.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySimpleGUI==4.60.4
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
python-vlc==3.0.16120
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.0
pywin32-ctypes==0.2.0
PyYAML==6.0
rapidfuzz==3.0.0
rarfile==4.0
requests==2.28.1
rich==13.3.4
scikit-image==0.19.3
scipy==1.7.3
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
streamlit==1.21.0
sympy==1.10.1
termcolor==2.3.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
tritonclient==2.33.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.2
urllib3==1.26.12
uvicorn==0.22.0
validators==0.20.0
vietocr==0.3.11
virtualenv==20.21.0
visualdl==2.5.2
watchdog==3.0.0
websockets==11.0.3
Werkzeug==2.2.3
x2paddle==1.4.1
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
